Applying recommended fix to use StringSubstitutor
(However a custom StringLookup class was needed, because StringSubstitutor normally would not support nested Maps as the valueMap)
