+++ components/inspectit-ocelot-config-doc-generator/build.gradle
@@ -29,6 +29,7 @@ dependencies {
             project(':inspectit-ocelot-config'),
             'ch.qos.logback:logback-classic:0.9.26',
             'org.apache.commons:commons-lang3:3.0',
+            'org.apache.commons:commons-text:1.9',
             'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:2.13.0',
             "com.fasterxml.jackson.datatype:jackson-datatype-jsr310",
             'commons-beanutils:commons-beanutils:1.9.4',
