+++ components/inspectit-ocelot-config-doc-generator/src/main/java/inspectit/ocelot/config/doc/generator/ConfigDocManager.java
@@ -2,6 +2,7 @@ package inspectit.ocelot.config.doc.generator;
 
 import inspectit.ocelot.config.doc.generator.docobjects.ConfigDocumentation;
 import inspectit.ocelot.config.doc.generator.docobjects.DocObjectGenerator;
+import parsing.ConfigParser;
 import rocks.inspectit.ocelot.config.model.InspectitConfig;
 
 public class ConfigDocManager {
