+++ components/inspectit-ocelot-config-doc-generator/src/main/java/parsing/StringSubstitutorNestedMap.java
@@ -1,0 +1,13 @@
+package parsing;
+
+import org.apache.commons.text.StringSubstitutor;
+
+import java.util.Map;
+
+public class StringSubstitutorNestedMap extends StringSubstitutor {
+
+    public StringSubstitutorNestedMap(Map<String, Object> valueMap){
+        this.setVariableResolver(new StringLookupNestedMap(valueMap));
+    }
+
+}
